refactor: Use zustand store instead of props in KakaoMap
