+++ app/(pages)/(main)/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect, useRef } from "react";
 import { useMainPageStore } from "@/stores/useMainPageStore";
 import ScrollToTopButton from "@/app/_ui/shared/ScrollToTopButton";
-import KakaoMap from "@/app/_ui/shared/KakaoMap";
+import KakaoMap from "./_ui/KakaoMap";
 import SearchBar from "./_ui/SearchBar";
 import NearbyFilter from "./_ui/NearbyFilter";
 import HospitalPharmacyTabs from "./_ui/HospitalPharmacyTabs";
@@ -69,10 +69,6 @@ export default function MainPage() {
     <div className="flex flex-col md:h-[calc(100vh-5rem)] md:flex-row">
       {/* Left: Kakao Map */}
       <KakaoMap
-        markerFilter={{
-          selectedMarker: activeTab,
-          onChange: (value) => setActiveTab(value),
-        }}
         hospitals={hospitals}
         pharmacies={pharmacies}
         ref={topButtonTriggerRef}
